Add BigQuery Storage API client (currently alpha) (#6696)
* Generate google-cloud-bigquery-storage library using synthtool. (#472)

* Add simple system test for BigQuery Storage API (#473)

* Add simple system test for BigQuery Storage API

For the system tests to run correctly, an unreleased version of api_core
is required. Updates the setup.py file to indicate the minimum version.

* Remove mock dependency and unnecessary imports.

* Use -e for local deps.

* Add manual client for BigQuery Storage (#474)

* Add manual client for BigQuery Storage.

The manual client adds a default project, which is used by the wrapped
create_read_session to make the project parameter optional.

A future purpose of the manual client is to override the read_rows
method to return an iterator that can reconnect with the correct offset,
much like Spanner's StreamedResultSet class. This work is not yet
complete.  I'd like feedback on the approach for the manual client
first.

* Use same client name as wrapped gapic class.

* Use subclass for BigQueryStorageClient

Also, change create_read_session project argument back to 'parent' to
better match the super class.

* Add unit tests.

* Remove default project from manual client.

* Use project_id fixture in system tests.

* Regenerate google-cloud-bigquery-storage (#475)

* Regenerate google-cloud-bigquery-storage

* Re-synthed

* Handwritten reader to reconnect on ReadRows streams (#477)

* WIP: reader to reconnect on ReadRows streams

* Add method to copy a StreamPosition.

* Add unit tests for hand-written reader module.

* Add docs for new hand-written module.

* Add fastavro to root intersphinx

* Add ServiceUnavailable to list of resumable exceptions.

Pull in .coveragerc file from upstream google-cloud-python.