feat: update synth to generate v1beta2, v1 endpoints for bigquerystorage
There's also some work to try to accomodate v1alpha2 in this PR, but
there exists further oddities that must be tackled before we can
fully add generation of the alpha client here.

Also, note that this PR does not include manual client modifications
such as streaming offset resumption that are present in the v1beta1
client.  Intent is to address further changes like that in subsequent
PRs.