Remove libcst dependency
libcst is only used for a one-off "fixup" script. It’s not necessary to always depend on it.

It caused an issue on one project as old libcst versions don't ship macOS M1 wheels, and it took time to figure this out.

(Also the instructions for using the fixup script were incorrect.)